Tentatively populate bundle versioning metadata
The meanings of and the distinctions between the different fields
are not obvious to me. I found a couple of resources worth noting, even
though I wouldn’t call them definitive.

* <https://eclecticlight.co/2021/06/19/explainer-version-and-build-numbers/>
* <https://eclecticlight.co/2017/10/14/which-version-was-that-how-to-check-which-version-youre-using/>

Anyway, I decided to use the number of commits in the current branch’s
history as the “Source Version” number. This can be obtained easily
with the Git command, `git rev-list --count HEAD`.